adjust java's long data type in number and binding map

+++ src/Parser/Expression/BinaryArithExpr.java
@@ -6,15 +6,15 @@ import java.util.Map;
 
 public record BinaryArithExpr(Expr left, String op, Expr right) implements Expr {
     @Override
-    public int eval(Map<String, Integer> bindings) throws EvalError {
-        int lv = left.eval(bindings);
-        int rv = right.eval(bindings);
+    public long eval(Map<String, Long> bindings) throws EvalError {
+        long lv = left.eval(bindings);
+        long rv = right.eval(bindings);
         if (op.equals("+")) return lv + rv;
         if (op.equals("-")) return lv - rv;
         if (op.equals("*")) return lv * rv;
         if (op.equals("/")) return lv / rv;
         if (op.equals("%")) return lv % rv;
-        if (op.equals("^")) return (int) Math.pow(lv, rv);
+        if (op.equals("^")) return (long) Math.pow(lv, rv);
         throw new EvalError("unknown op -> " + op);
     }
 }
