+++ src/Parser/Expression/Expr.java
@@ -5,5 +5,5 @@ import Parser.EvalError;
 import java.util.Map;
 
 public interface Expr {
-    int eval(Map<String, Integer> bindings) throws EvalError;
+    long eval(Map<String, Long> bindings) throws EvalError;
 }
