+++ src/Parser/Statement/Statement.java
@@ -5,6 +5,6 @@ import Parser.EvalError;
 import java.util.Map;
 
 public interface Statement {
-    void eval(Map<String, Integer> bindings) throws EvalError;
+    void eval(Map<String, Long> bindings) throws EvalError;
     default void append(Statement s) {}
 }
